requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.20.2
Pillow>=8.2.0
scipy>=1.7.1
